Relocate db/ & remove relative imports
