+++ packages/core/src/app/common/index.ts
@@ -1,3 +1,2 @@
 export * from './eth'
-export * from './db'
 export * from './utils'
