+++ packages/core/src/app/index.ts
@@ -1,3 +1,4 @@
 export * from './client'
 export * from './common'
 export * from './transport'
+export * from './db'
